make sure number casted.

+++ src/xlsx/read_sheet.rs
@@ -106,7 +106,7 @@ fn number_to_date(src: &String) -> DateTime<Utc> {
     let timestamp = ((((num - num.floor()) * 86400.0) as f64).round()) as i64;
     let hms = DateTime::from_timestamp(timestamp, 0).expect("unable to get datetime");
     let spreadsheet_epoch = Utc.with_ymd_and_hms(1900, 1, 1, 0, 0, 0).unwrap();
-    let duration = Duration::days(num.floor() -2) + Duration::hours(hms.hour().into()) + Duration::minutes(hms.minute().into()) + Duration::seconds(hms.second().into());
+    let duration = Duration::days(num.floor() as i64 -2) + Duration::hours(hms.hour().into()) + Duration::minutes(hms.minute().into()) + Duration::seconds(hms.second().into());
 
 
     let dt = (spreadsheet_epoch + duration);
